refactor: update to sveltkit style paths
This is where these would be if this were a sveltekit/vite
project that was started at the time of this commit. This is
prep for potentially switching from rollup builds to svelte.
to stay inside the svelte ecosystem.
