+++ src/lib/EventLog.svelte
